refactor(content): split telegram tgmbase service components
Closes #245
